move problem transformation to separate module
